Simplify OperationInternal
